test(selection): 마커 제거 스위트를 순수 함수 계약으로 분리한다 (#118)
앞 커밋이 가드를 더하며 HwpSelectionGeometryAttributedTests 본문이 319줄이 돼
swiftlint type_body_length 경고 구간(300)에 들어갔다 — 저장소가 문서화한
대응대로 관심사 경계에서 나눈다.

경계는 `strippingControlMarkerRuns`를 **직접** 부르는 순수 함수 계약(마커 전용
속성 소멸·감싼 링크 범위·선형 복잡도)과, 문서를 세워 `attributedText`가 그
결과로 무엇을 조립하는지 보는 계약이다. 앞엣것 3종이 새 스위트로 간다.

부수로 testLeadingMarkerOnlyFragmentDoesNotCrashNewlineAttributes의 낡은 주석을
지운다 — 그 경로는 이제 원본 마커 run에서 스타일을 읽어 `tail.length` 가드에
닿지 않는다.
